get_redis_master function moved to backends.py
